added in header content and adjusted colors
I added in the content for the upper portion of the header, still have to do the bottom portion. I did most of the styling for the header so I should just need to add in the HTML and do some minor adjustments. I also redid all the colors so it looks more presentable. I used the stone pallete from TailwindCSS.